feat: Set up checkCards array
This is an array that will hold any cards that are currently being
"checked" (i.e. they have not already been matched).

This will allow me to limit the player to two clicks, because if this
array has two cards in it, I'm going to disable the card click
event listener.