fix: Reset while wrong match open won't make error
Before, if you tried to reset the board while two wrong-match cards
were still open, it would throw an error because the hideAndRemove
method is being called asynchronously, and the cards were getting
removed by the reset action before they could get acted on by the
hideAndRemove function. This meant that hideAndRemove was trying to
act on undefined variables, because the cards were already removed
from the openCards array they were in by the reset function.

This change makes it so hideAndRemove only runs on the cars if
they exist in the openCards array.